survey_project/requirements.txt
asgiref==3.4.1
Django==2.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
